The begining

+++ WiseQueue.Domain/WiseQueue.Domain.Server/Management/ServerManager.cs
@@ -1,11 +1,14 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Common.Core.Logging;
 using WiseQueue.Core.Common.DataContexts;
 using WiseQueue.Core.Common.Management.Implementation;
-using WiseQueue.Core.Common.Models.Servers;
+using WiseQueue.Core.Common.Management.TaskManagment;
 using WiseQueue.Core.Server.Management;
+using WiseQueue.Domain.Common.Models;
+using WiseQueue.Domain.Common.Models.Servers;
 
 namespace WiseQueue.Domain.Server.Management
 {
@@ -18,20 +21,33 @@ namespace WiseQueue.Domain.Server.Management
         /// </summary>
         private readonly IServerDataContext serverDataContext;
 
+        /// <summary>
+        /// The <see cref="IQueueDataContext"/> instance.
+        /// </summary>
+        private readonly IQueueDataContext queueDataContext;
+
+        private readonly ITaskManager taskManager;
+
         private readonly TimeSpan heartbeatLifetime;
         private readonly TimeSpan sleepTime;
 
         private readonly CancellationTokenSource heartbetCancelationCancellationTokenSource;
         private readonly Task serverHeartbetTask;
 
+        private readonly ServerModel serverModel;
+        private readonly List<QueueModel> queues;
+
         #endregion
-        
+
         #region Properties...
 
         /// <summary>
         /// The server identifier.
         /// </summary>
-        public Int64 ServerId { get; private set; }
+        public Int64 ServerId
+        {
+            get { return serverModel.Id; }
+        }
 
         #endregion
 
@@ -41,14 +57,20 @@ namespace WiseQueue.Domain.Server.Management
         /// <param name="serverDataContext">The <see cref="IServerDataContext"/> instance.</param>
         /// <param name="loggerFactory">The <see cref="ICommonLoggerFactory"/> instance.</param>
         /// <exception cref="ArgumentNullException"><paramref name="serverDataContext"/> is <see langword="null"/></exception>
-        public ServerManager(IServerDataContext serverDataContext, ICommonLoggerFactory loggerFactory)
+        public ServerManager(IServerDataContext serverDataContext, IQueueDataContext queueDataContext, ITaskManager taskManager, ICommonLoggerFactory loggerFactory)
             : base("Server Manager", loggerFactory)
         {
             if (serverDataContext == null)
                 throw new ArgumentNullException(nameof(serverDataContext));
+            if (queueDataContext == null)
+                throw new ArgumentNullException(nameof(queueDataContext));
+            if (taskManager == null)
+                throw new ArgumentNullException(nameof(taskManager));
 
 
             this.serverDataContext = serverDataContext;
+            this.queueDataContext = queueDataContext;
+            this.taskManager = taskManager;
             heartbeatLifetime = TimeSpan.FromSeconds(15); //Move to settings.
 
             //5 seconds for connection to the SQL.
@@ -57,6 +79,16 @@ namespace WiseQueue.Domain.Server.Management
             heartbetCancelationCancellationTokenSource = new CancellationTokenSource();
             CancellationToken cancellationToken = heartbetCancelationCancellationTokenSource.Token;
             serverHeartbetTask = new Task(OnHeartbet, cancellationToken);
+
+            string serverName = Guid.NewGuid().ToString(); //TODO: Server's name should be more informative than Guid :)
+            string description = "This is a description"; //TODO: Server's description should be more informative than current one.
+            serverModel = new ServerModel(serverName, description, heartbeatLifetime);
+
+            //TODO: Populate queues from the configuration.
+            queues = new List<QueueModel>
+            {
+                new QueueModel("default", "This is a default queue.")
+            };
         }
 
         #region Implementation of IStartStoppable
@@ -66,13 +98,6 @@ namespace WiseQueue.Domain.Server.Management
         /// </summary>
         public void Start()
         {
-            string serverName = Guid.NewGuid().ToString(); //TODO: Server's name should be more informative than Guid :)
-            string description = "This is a description";
-            //TODO: Server's description should be more informative than current one.
-
-            ServerModel serverModel = new ServerModel(serverName, description, heartbeatLifetime);
-            ServerId = serverDataContext.InsertServer(serverModel);
-
             serverHeartbetTask.Start();
         }
 
@@ -91,25 +116,79 @@ namespace WiseQueue.Domain.Server.Management
         private void OnHeartbet(object state)
         {
             CancellationToken cancellationToken = (CancellationToken) state;
+            bool serverIsOnline = false;
 
             while (cancellationToken.IsCancellationRequested == false)
             {
-                logger.WriteTrace("Sending heartbeat for server id = {0}...", ServerId);
-                ServerHeartbeatModel serverHeartbeatModel = new ServerHeartbeatModel(ServerId, heartbeatLifetime);
-                serverDataContext.SendHeartbeat(serverHeartbeatModel);
-                logger.WriteTrace("The heartbeat has been sent.");
+                try
+                {
+                    if (serverIsOnline)
+                    {
+                        logger.WriteDebug("Sending heartbeat for server id = {0}...", ServerId);
+                        ServerHeartbeatModel serverHeartbeatModel = new ServerHeartbeatModel(ServerId, heartbeatLifetime);
+                        serverDataContext.SendHeartbeat(serverHeartbeatModel);
+                        logger.WriteDebug("The heartbeat has been sent.");
 
-                //Find servers that have been expired and delete them.
-                logger.WriteTrace("Finding and deleting servers that have been expired...");
-                int serverCount = serverDataContext.DeleteExpiredServers();
-                if (serverCount > 0)
-                    logger.WriteTrace("There were(was) {0} servers that have(has) been expired. They were(was) deleted.",
-                        serverCount);
-                else
-                    logger.WriteTrace("There was no any expired servers.");
+                        //Find servers that have been expired and delete them.
+                        logger.WriteDebug("Finding and deleting servers that have been expired...");
+                        int serverCount = serverDataContext.DeleteExpiredServers();
+                        if (serverCount > 0)
+                            logger.WriteDebug("There were(was) {0} servers that have(has) been expired. They were(was) deleted.", serverCount);
+                        else
+                            logger.WriteDebug("There was no any expired servers.");
+
+                        //Execute TaskManager.
+                        logger.WriteDebug("Executing task manager...");
+                        taskManager.Execute();
+                        logger.WriteDebug("The TaskManager has been executed.");
+                    }
+                    else
+                    {
+                        logger.WriteDebug("Preparing server for on-line...");
+
+                        logger.WriteTrace("Inserting information about server...");
+                        Int64 serverId = serverDataContext.InsertServer(serverModel);
+                        logger.WriteTrace("Inserting information about queues...");
+                        foreach (QueueModel queueModel in queues)
+                        {
+                            queueDataContext.InsertQueue(queueModel);
+                        }                        
+                        logger.WriteDebug("The server is on-line.");
+
+                        serverIsOnline = true;
+                        taskManager.SetOnlineStatus(true, serverId);
+                    }
+                }
+                catch (Exception ex)
+                {
+                    logger.WriteError(ex, "The exception has been raised during heartbeat.");
+                    serverIsOnline = false;
+                    taskManager.SetOnlineStatus(false);
+                }
 
                 cancellationToken.WaitHandle.WaitOne(sleepTime);
             }            
+        }
+
+
+        /// <summary>
+        /// Get default queue.
+        /// </summary>
+        /// <returns>The queue.</returns>
+        public QueueModel GetDefaultQueue()
+        {
+            QueueModel defaultQueue = queueDataContext.GetQueueByName("default");
+
+            return defaultQueue;
+        }
+
+        /// <summary>
+        /// Get all queues that available for this server,
+        /// </summary>
+        /// <returns>List of queues.</returns>
+        public IReadOnlyCollection<QueueModel> GetAvailableQueues()
+        {
+            throw new NotImplementedException();
         }
     }
 }
